@@ -19,8 +19,8 @@
 package com.gigigo.orchextrasdk.demo.ui.triggerlog.adapter;
 
 import android.content.Context;
-import android.support.annotation.DrawableRes;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.DrawableRes;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
