@@ -18,7 +18,7 @@
 
 package com.gigigo.orchextrasdk.demo.ui.triggerlog.filter.adapter;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
