@@ -21,8 +21,8 @@ package com.gigigo.orchextra.core.domain.actions.actionexecutors.webview
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.Toolbar
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.widget.Toolbar
 import com.gigigo.orchextra.core.R
 import kotlinx.android.synthetic.main.ox_activity_web_view.oxWebView
 import java.net.URI
