@@ -23,9 +23,9 @@ import android.content.Context
 import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Bundle
-import android.support.v4.app.ActivityCompat
-import android.support.v4.content.ContextCompat
-import android.support.v7.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import androidx.appcompat.app.AppCompatActivity
 import com.gigigo.orchextra.core.R
 import com.gigigo.orchextra.core.domain.entities.Error.Companion.PERMISSION_ERROR
 import com.gigigo.orchextra.core.domain.entities.Error.Companion.PERMISSION_RATIONALE_ERROR
