@@ -19,7 +19,7 @@
 package com.gigigo.orchextrasdk.demo.utils.widget;
 
 import android.content.Context;
-import android.support.design.widget.TextInputLayout;
+import com.google.android.material.textfield.TextInputLayout;
 import android.util.AttributeSet;
 import android.widget.EditText;
 import android.widget.LinearLayout;
