@@ -18,7 +18,6 @@
 
 package com.gigigo.orchextra.core.domain.entities
 
-import android.os.Parcel
 import android.os.Parcelable
 import com.gigigo.orchextra.core.domain.entities.TriggerType.BEACON
 import com.gigigo.orchextra.core.domain.entities.TriggerType.BEACON_REGION
@@ -26,22 +25,24 @@ import com.gigigo.orchextra.core.domain.entities.TriggerType.EDDYSTONE
 import com.gigigo.orchextra.core.domain.entities.TriggerType.EDDYSTONE_REGION
 import com.gigigo.orchextra.core.domain.entities.TriggerType.GEOFENCE
 import com.gigigo.orchextra.core.domain.entities.TriggerType.VOID
+import kotlinx.android.parcel.Parcelize
 
 
 enum class TriggerType {
-  BEACON,
-  BEACON_REGION,
-  EDDYSTONE,
-  EDDYSTONE_REGION,
-  GEOFENCE,
-  QR,
-  BARCODE,
-  IMAGE_RECOGNITION,
-  VOID;
+    BEACON,
+    BEACON_REGION,
+    EDDYSTONE,
+    EDDYSTONE_REGION,
+    GEOFENCE,
+    QR,
+    BARCODE,
+    IMAGE_RECOGNITION,
+    VOID;
 
-  infix fun withValue(value: String) = Trigger(this, value)
+    infix fun withValue(value: String) = Trigger(this, value)
 }
 
+@Parcelize
 data class Trigger constructor(
     val type: TriggerType,
     val value: String,
@@ -56,60 +57,28 @@ data class Trigger constructor(
     val detectedTime: Long = System.currentTimeMillis()
 ) : Parcelable {
 
-  constructor(source: Parcel) : this(
-      TriggerType.values()[source.readInt()],
-      source.readString(),
-      source.readValue(Double::class.java.classLoader) as Double?,
-      source.readValue(Double::class.java.classLoader) as Double?,
-      source.readString(),
-      source.readString(),
-      source.readString(),
-      source.readValue(Float::class.java.classLoader) as Float?,
-      source.readValue(Long::class.java.classLoader) as Long?,
-      source.readValue(Long::class.java.classLoader) as Long?
-  )
+    fun isVoid(): Boolean = type == VOID
 
-  fun isVoid(): Boolean = type == VOID
+    fun isBackgroundTrigger(): Boolean = (type == BEACON
+            || type == BEACON_REGION
+            || type == EDDYSTONE
+            || type == EDDYSTONE_REGION
+            || type == GEOFENCE)
 
-  fun isBackgroundTrigger(): Boolean = (type == BEACON
-      || type == BEACON_REGION
-      || type == EDDYSTONE
-      || type == EDDYSTONE_REGION
-      || type == GEOFENCE)
-
-  override fun describeContents() = 0
-
-  override fun writeToParcel(dest: Parcel, flags: Int) = with(dest) {
-    writeInt(type.ordinal)
-    writeString(value)
-    writeValue(lat)
-    writeValue(lng)
-    writeString(event)
-    writeString(phoneStatus)
-    writeString(distance)
-    writeValue(temperature)
-    writeValue(battery)
-    writeValue(uptime)
-  }
-
-  override fun equals(other: Any?): Boolean {
-    if (this === other) return true
-    if (javaClass != other?.javaClass) return false
-    other as Trigger
-
-    if (type != other.type) return false
-    if (value != other.value) return false
-    if (event != other.event) return false
-    if (distance != other.distance) return false
-    return true
-  }
-
-  companion object {
-    @JvmField
-    @Suppress("unused")
-    val CREATOR: Parcelable.Creator<Trigger> = object : Parcelable.Creator<Trigger> {
-      override fun createFromParcel(source: Parcel): Trigger = Trigger(source)
-      override fun newArray(size: Int): Array<Trigger?> = arrayOfNulls(size)
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (javaClass != other?.javaClass) return false
+        (other as? Trigger).let {
+            it ?: return false
+            if (type != it.type) return false
+            if (value != it.value) return false
+            if (event != it.event) return false
+            if (distance != it.distance) return false
+        }
+        return true
     }
-  }
+
+    override fun hashCode(): Int {
+        return super.hashCode()
+    }
 }
