@@ -31,91 +31,92 @@ import com.gigigo.orchextra.core.data.datasources.network.models.ApiDeviceInfo
 import com.gigigo.orchextra.core.data.datasources.network.models.ApiNotificationPush
 import com.gigigo.orchextra.core.data.datasources.network.models.ApiOxDevice
 import com.google.firebase.iid.FirebaseInstanceId
-import java.lang.Exception
-import java.util.Locale
-import java.util.TimeZone
-import java.util.UUID
+import java.util.*
 
 
 private const val TAG = "ContextExtensions"
 
 fun Context.getBaseApiOxDevice(anonymous: Boolean): ApiOxDevice = with(this) {
 
-  val wifiManager = applicationContext.getSystemService(Context.WIFI_SERVICE) as WifiManager
-  val bluetoothAdapter = BluetoothAdapter.getDefaultAdapter()
+    val wifiManager = applicationContext.getSystemService(Context.WIFI_SERVICE) as WifiManager
+    val bluetoothAdapter = BluetoothAdapter.getDefaultAdapter()
 
-  return ApiOxDevice(
-      instanceId = getIdToken(anonymous),
-      secureId = if (anonymous) {
-        "Anonymous"
-      } else {
-        Settings.Secure.getString(contentResolver, Settings.Secure.ANDROID_ID)
-      },
-      serialNumber = if (anonymous) {
-        "Anonymous"
-      } else {
-        Build.SERIAL
-      },
-      bluetoothMacAddress = if (anonymous) {
-        "Anonymous"
-      } else {
-        bluetoothAdapter?.address
-      },
-      wifiMacAddress = if (anonymous) {
-        "Anonymous"
-      } else {
-        wifiManager?.connectionInfo?.macAddress
-      },
-      clientApp = getApiClientApp(),
-      notificationPush = ApiNotificationPush(token = getFirebaseToken()),
-      device = getApiDeviceInfo(anonymous))
+    return ApiOxDevice(
+        instanceId = getIdToken(anonymous),
+        secureId = if (anonymous) {
+            "Anonymous"
+        } else {
+            Settings.Secure.getString(contentResolver, Settings.Secure.ANDROID_ID)
+        },
+        serialNumber = if (anonymous) {
+            "Anonymous"
+        } else {
+            Build.SERIAL
+        },
+        bluetoothMacAddress = if (anonymous) {
+            "Anonymous"
+        } else {
+            bluetoothAdapter?.address
+        },
+        wifiMacAddress = if (anonymous) {
+            "Anonymous"
+        } else {
+            wifiManager?.connectionInfo?.macAddress
+        },
+        clientApp = getApiClientApp(),
+        notificationPush = ApiNotificationPush(token = getFirebaseToken()),
+        device = getApiDeviceInfo(anonymous)
+    )
 }
 
 private fun Context.getApiClientApp(): ApiClientApp = with(this) {
-  val packageInfo = applicationContext.packageManager.getPackageInfo(packageName, 0)
+    val packageInfo = applicationContext.packageManager.getPackageInfo(packageName, 0)
 
-  return ApiClientApp(
-      bundleId = packageName,
-      buildVersion = packageInfo.versionCode.toString(),
-      appVersion = packageInfo.versionName,
-      sdkVersion = "-",
-      sdkDevice = "-")
+    return ApiClientApp(
+        bundleId = packageName,
+        buildVersion = packageInfo.versionCode.toString(),
+        appVersion = packageInfo.versionName,
+        sdkVersion = "-",
+        sdkDevice = "-"
+    )
 }
 
 private fun Context.getApiDeviceInfo(anonymous: Boolean): ApiDeviceInfo = with(this) {
-  val timeZone = TimeZone.getDefault()
-  val handset = if (Build.MODEL.startsWith(Build.MANUFACTURER)) {
-    Build.MODEL.capitalize()
-  } else {
-    "${Build.MANUFACTURER.capitalize()} ${Build.MODEL}"
-  }
+    val timeZone = TimeZone.getDefault()
+    val handset = if (Build.MODEL.startsWith(Build.MANUFACTURER)) {
+        Build.MODEL.capitalize()
+    } else {
+        "${Build.MANUFACTURER.capitalize()} ${Build.MODEL}"
+    }
 
-  return ApiDeviceInfo(
-      timeZone = "${timeZone.getDisplayName(false, TimeZone.SHORT)} ${timeZone.id}",
-      osVersion = if (anonymous) {
-        "Anonymous"
-      } else {
-        VERSION.SDK_INT.toString()
-      },
-      language = Locale.getDefault().toString(),
-      handset = if (anonymous) {
-        "Anonymous"
-      } else {
-        handset
-      },
-      type = "ANDROID")
+    return ApiDeviceInfo(
+        timeZone = "${timeZone.getDisplayName(false, TimeZone.SHORT)} ${timeZone.id}",
+        osVersion = if (anonymous) {
+            "Anonymous"
+        } else {
+            VERSION.SDK_INT.toString()
+        },
+        language = Locale.getDefault().toString(),
+        handset = if (anonymous) {
+            "Anonymous"
+        } else {
+            handset
+        },
+        type = "ANDROID"
+    )
 }
 
 private fun getIdToken(
-    anonymous: Boolean): String = if (getFirebaseToken().isEmpty() || anonymous) {
-  UUID.randomUUID().toString()
+    anonymous: Boolean
+): String = if (getFirebaseToken().isEmpty() || anonymous) {
+    UUID.randomUUID().toString()
 } else {
-  getFirebaseToken()
+    getFirebaseToken()
 }
 
 private fun getFirebaseToken(): String = try {
-  FirebaseInstanceId.getInstance().token ?: ""
+    FirebaseInstanceId.getInstance().token ?: ""
 } catch (e: Exception) {
-  Log.e(TAG, "getFirebaseToken()", e)
-  ""
+    Log.e(TAG, "getFirebaseToken()", e)
+    ""
 }
