@@ -18,10 +18,11 @@
 
 package com.gigigo.orchextra.indoorpositioning.domain.models
 
-import android.os.Parcel
 import android.os.Parcelable
-import java.util.Date
+import kotlinx.android.parcel.Parcelize
+import java.util.*
 
+@Parcelize
 data class OxBeacon(
     var hashcode: Int = 0, // hashcode()
     var beaconType: Int = -1, // Eddystone, altBeacon, iBeacon
@@ -44,89 +45,29 @@ data class OxBeacon(
     var temperature: Float = 0F,
     var pduCount: Long = 0,
     var uptime: Long = 0,
-    val lastDetection: Date) : Parcelable {
+    val lastDetection: Date
+) : Parcelable {
 
-  fun getTemperatureInCelsius(): Float {
-    val tmp = temperature / 256F
+    fun getTemperatureInCelsius(): Float {
+        val tmp = temperature / 256F
 
-    if (tmp == (1 shl 7).toFloat()) { // 0x8000
-      return 0F
+        if (tmp == (1 shl 7).toFloat()) { // 0x8000
+            return 0F
+        }
+        return if (tmp > (1 shl 7)) tmp - (1 shl 8) else tmp
     }
-    return if (tmp > (1 shl 7)) tmp - (1 shl 8) else tmp
-  }
 
-  fun getDistanceQualifier(): String = when {
-    distance < 0.5 -> "immediate"
-    distance < 5 -> "near"
-    distance < 20 -> "far"
-    else -> "unknown"
-  }
-
-  constructor(source: Parcel) : this(
-      source.readInt(),
-      source.readInt(),
-      source.readString(),
-      source.readString(),
-      source.readInt(),
-      source.readInt(),
-      source.readInt(),
-      source.readInt(),
-      source.readDouble(),
-      source.readLong(),
-      source.readLong(),
-      source.readInt(),
-      source.readString(),
-      source.readString(),
-      source.readString(),
-      1 == source.readInt(),
-      source.readLong(),
-      source.readLong(),
-      source.readFloat(),
-      source.readLong(),
-      source.readLong(),
-      source.readSerializable() as Date
-  )
-
-  override fun describeContents() = 0
-
-  override fun writeToParcel(dest: Parcel, flags: Int) = with(dest) {
-    writeInt(hashcode)
-    writeInt(beaconType)
-    writeString(beaconAddress)
-    writeString(uuid)
-    writeInt(major)
-    writeInt(minor)
-    writeInt(txPower)
-    writeInt(rssi)
-    writeDouble(distance)
-    writeLong(lastSeen)
-    writeLong(lastMinuteSeen)
-    writeInt(manufacturer)
-    writeString(url)
-    writeString(namespaceId)
-    writeString(instanceId)
-    writeInt((if (hasTelemetryData) 1 else 0))
-    writeLong(telemetryVersion)
-    writeLong(batteryMilliVolts)
-    writeFloat(temperature)
-    writeLong(pduCount)
-    writeLong(uptime)
-    writeSerializable(lastDetection)
-  }
-
-  companion object {
-    const val TYPE_EDDYSTONE_UID = 0
-
-    const val TYPE_EDDYSTONE_URL = 1
-
-    const val TYPE_ALTBEACON = 2
-
-    const val TYPE_IBEACON = 3
-
-    @JvmField
-    val CREATOR: Parcelable.Creator<OxBeacon> = object : Parcelable.Creator<OxBeacon> {
-      override fun createFromParcel(source: Parcel): OxBeacon = OxBeacon(source)
-      override fun newArray(size: Int): Array<OxBeacon?> = arrayOfNulls(size)
+    fun getDistanceQualifier(): String = when {
+        distance < 0.5 -> "immediate"
+        distance < 5 -> "near"
+        distance < 20 -> "far"
+        else -> "unknown"
     }
-  }
+
+    companion object {
+        const val TYPE_EDDYSTONE_UID = 0
+        const val TYPE_EDDYSTONE_URL = 1
+        const val TYPE_ALTBEACON = 2
+        const val TYPE_IBEACON = 3
+    }
 }
